requirements.txt
flask
openai
pytesseract
requests
Pillow
